// app/src/generated/models/Al_reviewinstancesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Al_reviewinstancesal_reviewstatus = {
  120910210: 'Assigned',
  120910211: 'Review In Progress',
  120910212: 'Submitted'
} as const;
export type Al_reviewinstancesal_reviewstatus = keyof typeof Al_reviewinstancesal_reviewstatus;
export const Al_reviewinstancesal_reviewtype = {
  120910200: 'Tax',
  120910201: 'AQS'
} as const;
export type Al_reviewinstancesal_reviewtype = keyof typeof Al_reviewinstancesal_reviewtype;
export const Al_reviewinstancesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Al_reviewinstancesstatecode = keyof typeof Al_reviewinstancesstatecode;
export const Al_reviewinstancesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Al_reviewinstancesstatuscode = keyof typeof Al_reviewinstancesstatuscode;

export interface Al_reviewinstancesBase {
  "al_checklistversionid@odata.bind": string;
  al_name: string;
  "al_outcomecaseid@odata.bind"?: string;
  al_reviewinstancecode: string;
  al_reviewinstanceid: string;
  al_reviewstatus: Al_reviewinstancesal_reviewstatus;
  al_reviewtype: Al_reviewinstancesal_reviewtype;
  al_sequence: number;
  al_startedon?: string;
  al_submittedon?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Al_reviewinstancesstatecode;
  statuscode?: Al_reviewinstancesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Al_reviewinstances extends Al_reviewinstancesBase {
  al_checklistversionidname?: string;
  al_outcomecaseidname?: string;
  al_reviewstatusname?: string;
  al_reviewtypename?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  al_checklistversionid?: object;
  _al_checklistversionid_value?: string;
  al_outcomecaseid?: object;
  _al_outcomecaseid_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}